fix: Fixing bug

+++ Application/Meetups/Commands/DeleteMeetup.cs
@@ -2,18 +2,24 @@ using Persistence;
 
 namespace Application.Meetups.Commands
 {
-    public class DeleteMeetupService(AppDbContext context)
+    public class DeleteMeetupHandler(AppDbContext context)
     {
         private readonly AppDbContext _context = context;
 
-        public async Task DeleteMeetupAsync(string id, CancellationToken cancellationToken = default)
+        public async Task Handle(DeleteMeetupCommand command, CancellationToken cancellationToken = default)
         {
-            var meetup = await _context.Meetups
-                .FindAsync([id], cancellationToken) 
-                    ?? throw new Exception($"Meetup with ID {id} not found");
+            var meetup = await _context.Meetups.FindAsync(command.Id, cancellationToken);
+
+            if (meetup is null)
+                throw new KeyNotFoundException($"Meetup with ID '{command.Id}' not found.");
 
             _context.Meetups.Remove(meetup);
             await _context.SaveChangesAsync(cancellationToken);
         }
+    }
+
+    public class DeleteMeetupCommand
+    {
+        public string Id { get; set; } = default!;
     }
 }
